requirements.txt
Pillow
Django
django-registration-redux
